Revert "Revert "[aot_inductor] move CudaWrapperCodeGen into a separate file (#119448)""
This reverts commit 3ab08946d5052eaeda11d683d6a58e801a032755.

[ghstack-poisoned]
